refactor for db folders
