+++ app/api/ada/lib/storage/database/transactionModels/account/api/read.js
@@ -14,8 +14,8 @@ import type {
 } from '../tables';
 import type {
   TransactionRow,
-} from '../../primitives/tables';
-import { getRowIn, } from '../../utils';
+} from '../../../primitives/tables';
+import { getRowIn, } from '../../../utils';
 
 export class GetAccountingInputs {
   static ownTables = Object.freeze({
