+++ app/api/ada/lib/storage/database/transactionModels/multipart/tables.js
@@ -2,9 +2,9 @@
 
 import type { lf$schema$Builder } from 'lovefield';
 
-import type { DbTransaction, DbBlock, } from '../primitives/tables';
-import type { DbUtxoInputs, DbUtxoOutputs, } from '../utxoTransactions/tables';
-import type { DbAccountingInputs, DbAccountingOutputs, } from '../accountingTransactions/tables';
+import type { DbTransaction, DbBlock, } from '../../primitives/tables';
+import type { DbUtxoInputs, DbUtxoOutputs, } from '../utxo/tables';
+import type { DbAccountingInputs, DbAccountingOutputs, } from '../account/tables';
 
 export type DbTxIO = {|
   ...DbTransaction,
