+++ app/api/ada/lib/storage/database/walletTypes/bip44/api/utils.js
