+++ app/api/ada/lib/storage/models/PublicDeriver/interfaces.js
@@ -13,16 +13,16 @@ import { ConceptualWallet } from '../ConceptualWallet/index';
 
 import type {
   TreeInsert,
-} from '../../database/bip44/api/write';
+} from '../../database/walletTypes/bip44/api/write';
 import type {
   Bip44AddressRow,
   AccountingDerivationRow,
-} from '../../database/bip44/tables';
+} from '../../database/walletTypes/bip44/tables';
 
 import {
   GetUtxoTxOutputsWithTx,
-} from '../../database/utxoTransactions/api/read';
-import type { UtxoTxOutput } from '../../database/utxoTransactions/api/read';
+} from '../../database/transactionModels/utxo/api/read';
+import type { UtxoTxOutput } from '../../database/transactionModels/utxo/api/read';
 
 import type {
   AddressRow,
@@ -50,8 +50,8 @@ import {
 import {
   AddBip44Tree,
   ModifyDisplayCutoff,
-} from '../../database/bip44/api/write';
-import { GetBip44DerivationSpecific } from '../../database/bip44/api/read';
+} from '../../database/walletTypes/bip44/api/write';
+import { GetBip44DerivationSpecific } from '../../database/walletTypes/bip44/api/read';
 import { UpdateGet, GetOrAddAddress, } from '../../database/primitives/api/write';
 import type {
   FilterFunc,
