+++ app/api/ada/transactions/types.js
@@ -7,7 +7,7 @@ import { RustModule } from '../lib/cardanoCrypto/rustLoader';
 
 import type {
   DbTxIO,
-} from '../lib/storage/database/multipartTransactions/tables';
+} from '../lib/storage/database/transactionModels/multipart/tables';
 import type { DbBlock, } from '../lib/storage/database/primitives/tables';
 import type {
   Address, Value, Addressing,
